update due to lint findings
git-svn-id: https://svn.code.sf.net/p/tcnopen/trdp/trunk/trdp@441 3b5a3598-5f4e-4449-9e63-bd40438bfec0

+++ src/common/trdp_pdcom.c
@@ -598,7 +598,7 @@ TRDP_ERR_T trdp_pdCheck (
     if (packetSize < TRDP_MIN_PD_HEADER_SIZE ||
         packetSize > TRDP_MAX_PD_PACKET_SIZE)
     {
-        vos_printf(VOS_LOG_INFO, "PDframe size error (%u))\n", (UINT32) packetSize);
+        vos_printf(VOS_LOG_INFO, "PDframe size error (%u))\n", packetSize);
         err = TRDP_WIRE_ERR;
     }
